+++ src/main/java/lotto/io/OutputMessage.java
@@ -1,4 +1,4 @@
-package lotto.message;
+package lotto.io;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -6,6 +6,7 @@ import java.util.HashMap;
 import java.util.List;
 import lotto.data.Lotto;
 import lotto.domain.LottoResult;
+import lotto.message.LottoPrize;
 
 public enum OutputMessage {
     ASK_PURCHASE_AMOUNT("구입금액을 입력해 주세요."),
